Improvement: add auxiliary submission actions
